+++ core/model/task.go
@@ -1,12 +1,11 @@
 package model
 
 import (
-	log "github.com/cihub/seelog"
-	"github.com/infinitbyte/gopa/core/errors"
-	"github.com/infinitbyte/gopa/core/store"
-
 	"bytes"
 	"fmt"
+	log "github.com/cihub/seelog"
+	"github.com/infinitbyte/gopa/core/errors"
+	"github.com/infinitbyte/gopa/core/persist"
 	"github.com/infinitbyte/gopa/core/pipeline"
 	"github.com/infinitbyte/gopa/core/util"
 	"strconv"
@@ -44,8 +43,8 @@ func GetTaskStatusText(status TaskStatus) string {
 type Seed struct {
 	Url       string `storm:"index" json:"url,omitempty" gorm:"type:varchar(500)"` // the seed url may not cleaned, may miss the domain part, need reference to provide the complete url information
 	Reference string `json:"reference_url,omitempty"`
-	Depth     int    `storm:"index" json:"depth,omitempty"`
-	Breadth   int    `storm:"index" json:"breadth,omitempty"`
+	Depth     int    `storm:"index" json:"depth"`
+	Breadth   int    `storm:"index" json:"breadth"`
 }
 
 func (this Seed) Get(url string) Seed {
@@ -116,24 +115,24 @@ func NewTaskSeed(url, ref string, depth int, breadth int) Seed {
 
 type Task struct {
 	Seed
-	ID            string          `gorm:"not null;unique;primary_key" json:"id"`
-	Host          string          `gorm:"index" json:"-"`
-	Schema        string          `json:"schema,omitempty"`
-	OriginalUrl   string          `json:"original_url,omitempty"`
-	Phrase        pipeline.Phrase `gorm:"index" json:"phrase"`
-	Status        TaskStatus      `gorm:"index" json:"status"`
-	Message       string          `json:"-"`
-	CreateTime    *time.Time      `gorm:"index" json:"created,omitempty"`
-	UpdateTime    *time.Time      `gorm:"index" json:"updated,omitempty"`
-	LastFetchTime *time.Time      `gorm:"index" json:"last_fetch"`
-	LastCheckTime *time.Time      `gorm:"index" json:"last_check"`
-	NextCheckTime *time.Time      `gorm:"index" json:"next_check"`
+	ID          string          `gorm:"not null;unique;primary_key" json:"id" index:"id"`
+	Host        string          `gorm:"index" json:"host"`
+	Schema      string          `json:"schema,omitempty"`
+	OriginalUrl string          `json:"original_url,omitempty"`
+	Phrase      pipeline.Phrase `gorm:"index" json:"phrase"`
+	Status      TaskStatus      `gorm:"index" json:"status"`
+	Message     string          `json:"-"`
+	Created     *time.Time      `gorm:"index" json:"created,omitempty"`
+	Updated     *time.Time      `gorm:"index" json:"updated,omitempty"`
+	LastFetch   *time.Time      `gorm:"index" json:"last_fetch"`
+	LastCheck   *time.Time      `gorm:"index" json:"last_check"`
+	NextCheck   *time.Time      `gorm:"index" json:"next_check"`
 
-	SnapshotVersion    int        `json:"snapshot_version"`
-	SnapshotID         string     `json:"snapshot_id"`      //Last Snapshot's ID
-	SnapshotHash       string     `json:"snapshot_hash"`    //Last Snapshot's Hash
-	SnapshotSimHash    string     `json:"snapshot_simhash"` //Last Snapshot's Simhash
-	SnapshotCreateTime *time.Time `json:"snapshot_created"` //Last Snapshot's Simhash
+	SnapshotVersion int        `json:"snapshot_version"`
+	SnapshotID      string     `json:"snapshot_id"`      //Last Snapshot's ID
+	SnapshotHash    string     `json:"snapshot_hash"`    //Last Snapshot's Hash
+	SnapshotSimHash string     `json:"snapshot_simhash"` //Last Snapshot's Simhash
+	SnapshotCreated *time.Time `json:"snapshot_created"` //Last Snapshot's Simhash
 }
 
 func CreateTask(task *Task) error {
@@ -141,9 +140,9 @@ func CreateTask(task *Task) error {
 	time := time.Now().UTC()
 	task.ID = util.GetIncrementID("task")
 	task.Status = TaskCreated
-	task.CreateTime = &time
-	task.UpdateTime = &time
-	err := store.Save(task)
+	task.Created = &time
+	task.Updated = &time
+	err := persist.Save(task)
 	if err != nil {
 		log.Error(task.ID, ", ", err)
 	}
@@ -153,8 +152,8 @@ func CreateTask(task *Task) error {
 func UpdateTask(task *Task) {
 	log.Trace("start update crawler task, ", task.Url)
 	time := time.Now().UTC()
-	task.UpdateTime = &time
-	err := store.Update(task)
+	task.Updated = &time
+	err := persist.Update(task)
 	if err != nil {
 		panic(err)
 	}
@@ -163,7 +162,7 @@ func UpdateTask(task *Task) {
 func DeleteTask(id string) error {
 	log.Trace("start delete crawler task: ", id)
 	task := Task{ID: id}
-	err := store.Delete(&task)
+	err := persist.Delete(&task)
 	if err != nil {
 		log.Error(id, ", ", err)
 	}
@@ -173,11 +172,12 @@ func DeleteTask(id string) error {
 func GetTask(id string) (Task, error) {
 	log.Trace("start get seed: ", id)
 	task := Task{}
-	err := store.GetBy("id", id, &task)
+	task.ID = id
+	err := persist.Get(&task)
 	if err != nil {
 		log.Error(id, ", ", err)
 	}
-	if len(task.ID) == 0 || task.CreateTime == nil {
+	if len(task.ID) == 0 || task.Created == nil {
 		panic(errors.New("not found," + id))
 	}
 
@@ -187,7 +187,7 @@ func GetTask(id string) (Task, error) {
 func GetTaskByField(k, v string) (Task, error) {
 	log.Trace("start get seed: ", k, ", ", v)
 	task := Task{}
-	err := store.GetBy(k, v, &task)
+	err := persist.GetBy(k, v, &task)
 	if err != nil {
 		log.Error(k, ", ", err)
 	}
@@ -197,25 +197,38 @@ func GetTaskByField(k, v string) (Task, error) {
 func GetTaskList(from, size int, domain string) (int, []Task, error) {
 	log.Tracef("start get crawler tasks, %v-%v, %v", from, size, domain)
 	var tasks []Task
-	queryO := store.Query{Sort: "create_time desc", From: from, Size: size}
+	sort := []persist.Sort{}
+	sort = append(sort, persist.Sort{Field: "created", SortType: persist.DESC})
+	queryO := persist.Query{Sort: &sort, From: from, Size: size}
 	if len(domain) > 0 {
-		queryO.Conds = store.And(store.Eq("host", domain))
+		queryO.Conds = persist.And(persist.Eq("host", domain))
 	}
-	err, result := store.Search(&tasks, &queryO)
+	err, result := persist.Search(Task{}, &tasks, &queryO)
 	if err != nil {
-		log.Error(err)
+		panic(err)
 	}
+	if tasks == nil {
+		convertTask(result, &tasks)
+	}
+
 	return result.Total, tasks, err
 }
 
 func GetPendingNewFetchTasks() (int, []Task, error) {
 	log.Trace("start get all crawler tasks")
 	var tasks []Task
-	queryO := store.Query{Sort: "create_time desc", Conds: store.And(store.Eq("phrase", 1))}
-	err, result := store.Search(&tasks, &queryO)
+	sort := []persist.Sort{}
+	sort = append(sort, persist.Sort{Field: "created", SortType: persist.DESC})
+	queryO := persist.Query{Sort: &sort, Conds: persist.And(persist.Eq("phrase", 1))}
+	err, result := persist.Search(Task{}, &tasks, &queryO)
 	if err != nil {
 		log.Error(err)
 	}
+
+	if tasks == nil {
+		convertTask(result, &tasks)
+	}
+
 	return result.Total, tasks, err
 }
 
@@ -223,14 +236,33 @@ func GetPendingUpdateFetchTasks(offset *time.Time) (int, []Task, error) {
 	t := time.Now().UTC()
 	log.Tracef("start get all crawler tasks,last offset: %s,", offset.String())
 	var tasks []Task
-	queryO := store.Query{Sort: "create_time asc",
-		Conds: store.And(
-			store.Lt("next_check_time", t),
-			store.Gt("create_time", offset),
-			store.Eq("status", TaskFetchSuccess))}
-	err, result := store.Search(&tasks, &queryO)
+	sort := []persist.Sort{}
+	sort = append(sort, persist.Sort{Field: "created", SortType: persist.ASC})
+	queryO := persist.Query{Sort: &sort,
+		Conds: persist.And(
+			persist.Lt("next_check", t),
+			persist.Gt("created", offset),
+			persist.Eq("status", TaskFetchSuccess)),
+		From: 0, Size: 100,
+	}
+	err, result := persist.Search(Task{}, &tasks, &queryO)
 	if err != nil {
 		log.Error(err)
 	}
+
+	if tasks == nil {
+		convertTask(result, &tasks)
+	}
+
 	return result.Total, tasks, err
+}
+
+func convertTask(result persist.Result, tasks *[]Task) {
+	t := result.Result.([]interface{})
+	for _, i := range t {
+		js := util.ToJson(i, false)
+		t := Task{}
+		util.FromJson(js, &t)
+		*tasks = append(*tasks, t)
+	}
 }