+++ CHANGES.md
@@ -17,6 +17,7 @@
 4. add proxy to fetch joint
 5. filter url before push to checker
 6. add rules config to url filter 
+7. support elasticsearch as database store
 
 ##### improvement
 1. multi instance support on local machine